Added a panel to show the status of the apt packages
This allows a user to include a status panel of what package updates are pending on an Apt Package
managed system such as Ubuntu.

Also included is the configuration file change to include a 'package_management' section that must
be configured for proper operation. (I was unable to get the panel to not display if the section was
missing) It is currently configured to be not enabled. Change the flag to 'true' to enable the panel.

While setup for Apt, this could easily be extended for yum, pacman, etc and I hope they would include
their configuration inside of 'package_management' as well.

Tested on: Ubuntu 20.04.2 LTS (GNU/Linux 5.4.0-72-generic x86_64)